Split heideltime class in two: refactor ambiguous values resolving.
